requirements.txt
fastapi
uvicorn
jinja2
aiofiles
python-periphery
opencv-python